dtype fixed

+++ main.py
@@ -9,9 +9,9 @@ if __name__ == "__main__":
     print("Вводите уравнения в формате 'a1 a2 ... am | b1', где ai - коэффициенты при переменных, bj - свободные члены")
     for i in range(n):
         *ai, _, bi = list(input().split())
-        a.append(list(map(float, ai)))
-        b.append([float(bi)])
-    A = np.array(a)
-    b = np.array(b)
+        a.append(list(map(int, ai)))
+        b.append([int(bi)])
+    A = np.array(a, dtype=np.int64)
+    b = np.array(b, dtype=np.int64)
     gaussian_elimination(A, b)
     exit = input("Нажмите Enter, чтобы закрыть программу")
